Modify classes for better OOP
